hydra-module/compressLogs: Avoid creating /root/r.
We really don't need to touch a file in the current working directory
to find files that are older than one month. Since findutils 4.3.3 there
is a -newerXY option which allows to specify timestamps directly (as
with `date --date`).

But even when using a reference file, it really causes confusion if
people look into /root and try to debug where that misterious "r" file
is coming from.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>