hydra-module: Use PermissionsStartOnly in init.
Oops, forgot to add this in f75509099a558ecfdd2f60a036c6e71de430920a.

This is necessary because we actually want to run the preStart script as
root (because it chmod/chowns stuff and also needs to create the
database using PostgreSQL's superuser) and the actual creation of the
database as user hydra.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>