Add a new column "channelAttr" to Jobsets table.
Needed in order to actually allow setting the channel paths in jobsets
and to actually deliver the right Nix expressions.

Unfortunately this wasn't as easy as I thought it will be, because
issuing a "make update-dbix" in the src/sql directory created the
following additional fields:

jobsets_2s
buildsteps_builds
buildsteps_propagatedfroms

All of them seem to be related to many2many relations, and I've used the
same version of DBIx::Class::Schema::Loader as the one generated before.

I've checked whether there were commits related to manual changes or
even fields that someone forgot to create but found nothing about it.

So I ran update-dbix with the unmodified SQL file and source tree,
committed the changes, add the column, commit this very commit, rebase
and throw away the previous commit.

That way I hope everything stays consistent.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>