release.nix: Add VM test for custom channels.
It's only a basic test to ensure that the given channel job is correctly
delivered when nix-channel --update is issued.

We're also checking whether another similarily structured project
affects the current channel to make sure we don't have a bug while
selecting the builds for the corresponding channel.

Also the goal is that the channel expression is independent of the whole
evaluation, so we also have channel-unrelated jobs that deliberately
fail while the channel does not, so the recipient still gets updates.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>