Implement new view for custom channels.
It's probably not very efficient if it comes to my use of the ORM, as
I'm not very familiar with Catalyst. But apart from that, it's basically
like this: If the channelAttr column is not null, we're going to deliver
a tarball of the first build output from the job specified by that
column. Otherwise, the behavior is like before.

As for the view itself, the reason why I'm using GNU Tar instead of
Archive::Tar is that the latter keeps the contents of the tarball in
memory, as pointed out at its FAQ:

http://search.cpan.org/~bingos/Archive-Tar-2.04/lib/Archive/Tar.pm#FAQ

Also, it's a bit difficult to do filename transformation with the Perl
module, so we would have to use things such as $tar->{_data}, which by
its underscore in its name doesn't sound like the author of the library
wants to keep the interface for it very stable accross versions.

However, when it comes to bzip2, we're still using IO::Compress::Bzip2,
because it doesn't seem have the same issue as Archive::Tar.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>