add default.nix and fix build
This commit adds a default.nix that is extracted from the build part
of release.nix. This makes it simpler to build hydra from source,
without having to remeber to specify the correct system, a simple
'nix-build' in the source root is enough.

The commit also changes the hydra build to use a specific nix commit
instead of using the nixUnstable from nixpkgs. This fixes a build error
because hydra currently does not compile with latest nix master.