release.nix: Make it easier to set up dev database
This is for setting up a local PostgreSQL instance inside the
inst/database directory.

I'm re-using this directory because I conveniently found it inside the
.gitignore file, so let's just use that instead of adding a new path to
the .gitignore file.

The functions are currently a bit fragmented and documentation is
lacking, but in order to spin up a fully working dev server, you now
just need to do the following:

  $ nix-shell
  $ ./bootstrap
  $ ./configure $configureFlags
  $ make
  $ setup-database
  $ hydra-server

Of course, we can still improve on this, but this is at least a good
base, which also makes sure that whenever the shell exits the database
server is stopped as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>