shell-hook: CamelCase internal functions/variables
I've used dashed function names to make the names a bit more natural to
users of shells. So let's use CamelCased functions and variable names
for all of the internal stuff to make it more unlikely that users
accidentally call those.

Of course, everybody is still free to call them and the world will not
explode. :-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>