shell-hook: Setting up env vars for setup-database
If we're running directly with the full setup-dev-env, the setupEnvVars
function is not yet called because the database doesn't exist yet.

So let's use setupEnvVars prior to initializing the database so that all
environment variables are properly set up.

This also fixes a circular dependency on setup-dev-env, because that was
what setupEnvVars was called before and we accidentally left it in.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>