shell-hook: Use setsid() to detach PostgreSQL
Without this, whenever we press CTRL+C within the nix-shell, the
database server will get SIGINT and will terminate.

That's obviously a very unexpected situation, so let's start PostgreSQL
using the setsid command from util-linux.

Obviously that command is only available on GNU/Linux, so we make sure
the setsid command is only used whenever it is available.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>