shell-hook: Implement semaphore for DBMS
If we're going to use multiple nix-shells we don't want to spin down the
development database every time another nix-shell is closed, nor do we
want to have multiple database servers running with the same data
directory (which fails anyway).

Now the database server is started whenever we enter the first nix-shell
and it's shutdown when the last nix-shell is closed.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>