Turn shell.nix into default.nix with working Nix
The current Nix master version doesn't build with Hydra and nixUnstable
doesn't work either.

This is a variation of @bennofs approach in #347, having a fixed and
known to work/build Nix version in default.nix.

While his approach moves the contents of the release.nix into the
default.nix, I'm doing it the other way around, so that we still have
release.nix solely for Hydra with the jobset input attributes
parameterizable.

To get people started as quickly as possible, the default.nix also comes
without input attributes that deal with release handling.

Having a default.nix, we no longer need shell.nix, because we can simply
test for IN_NIX_SHELL (or in this case lib.inNixShell, which does the
same at the moment).

Credits go to @bennofs for stealing parts of his comments and ripping
off his idea, thanks :-)

Signed-off-by: aszlig <aszlig@redmoonstudios.org>