Replace table jobset_input_alts by a JSONB column
We could have used the classic entity-attribute-value (EAV) model
instead but that would involve lots of self-joins and we'd also need to
put every single property value into a series of text fields.

If we want to have nesting, like we have for the "attrs" value for the
"build" and "sysbuild" types we'd have to introduce even more madness by
even creating yet another EAV table.

Of course, we could avoid the nesting by serializing the stuff into the
text fields, but serialization would be something like JSON anyway and
we have a field type for that since PostgreSQL 9.3 and even a JSONB type
in version 9.4 which is even indexable.

The really nasty part here was getting every jobset input type into JSON
here, because during a schema upgrade we can't rely on PL/Perl being
available and everything outside seems to be hackish by modifying the
schema updater code to run a Perl function just for upgrade 51 (which
also needs to be retained).

Right now, this only updates the database and the DBIx schema (also with
deflation/inflation of properties from/to JSON) but not the code, which
is still using jobset_input_alts and will obviously fail now.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Issue: #279