plugins: Handle properties for supportedInputTypes
Switches the supportedInputTypes sub to use the new property system.

The documentation in Hydra/Plugin.pm is not yet finished because the
structure of the properties might change while I'm working on this.

But in the end we now can specify input properties in a declarative way
rather than parsing a single string field.

Editing and displaying these properties currently still throws an error.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>