jobset: Fix rendering properties in configuration
We now have a hashmap called inputTypeNames we give to the template, so
that we can simply pass this to renderSelection for the root type.

For the properties themselves, we can currently render singleton types
and multi-value types, including attrsets. However, we still don't
handle optional fields yet.

The UI design of these properties still needs to be revamped as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>