jobset: Allow to use HTTP OPTIONS method
When usisng OPTIONS on a jobset, we can get the declarative
specification of the properties. This is not only useful for our
internal UI, but is also useful for external applications, because they
can validate these properties before sending them to our backend.

Ideally, we would not only expose OPTIONS for just the properties but
also for all the other fields of a jobset as well. But for now, let's
focus on the properties only and expand on this later.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>