edit-jobset: Implement basic property editing
This is mainly a lot of JavaScript, because we don't know beforehand how
many properties there are and also we have a property type for
"attrset", which consists of arbitrary fields.

Getting back the values of all those form fields, which could be
arbitrary and nested is one the trickiest part here.

The way this works is that the specification of these properties is
retrieved into a static variable using an OPTIONS request on the jobset
URL. Afterwards the specification is used for generating all the DOM
nodes and the discharger functions.

When the final form is submitted, the discharger functions are called
for each property and they return the current values of the form field,
which in turn gets serialized into JSON.

However, the implementation is very basic and currently doesn't account
for recursive attrset properties nor does it handle optional fields or
help texts.

Attrset properties currently are lacking an "add attribute" button to
create a new key + value, so it's pretty useless right now unless you
edit an existing attrset.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>