Jobset: Make validation of "eval" a property field
We now have a "validate" key for the property specification, which
allows custom functions to check whether the input is valid or not.

Right now, this isn't checked on the PUT method of jobsets yet.

As the validation function can't be serialized, we also need to make
sure it's stripped off during an OPTIONS request on jobsets.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>