Jobset: Validate properties on PUT request
This replaces the previous checkInputValue sub with a new
validateProperty sub that checks every property type for its validity
and also calls the "validate" function if it exists.

We now get a proper error message for both the API users and the web UI
if the submitted input is wrong.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>