Implement handling properties in fetchInput
This is on the side of the evaluator, where we didn't handle the new
properties so far.

The code removed is were the ugliest parts of the previous
implementation, because they were trying to parse all those attributes
using string matching.

While working on this change, what particularly struck me was this:

    if ($exprType eq "guile") {
        if ($value eq "true") {
            $result = "#t";
        } else {
            $result = "#f";
        }
        $result = $value;
    }

All of this is more or less a no-op, because $result gets overwriten by
the original $value.

So I've changed it to return "#t"/"#f" for Guile now.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @civodul