Fix tests to handle the new input property system
I've ran all the tests currently exposed by release.nix and of course
also the tests that are run in build.$system.

All of the tests now succeed and a "git grep -i alts" only gives back
src/lib/Hydra/Controller/API.pm as the last step that's left to convert
to properties.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>