Schema/JobsetInputs: Bring back json_hint
I've removed this in e0a2656, thinking it wasn't needed because it was
only referring to jobset_input_alts. Unfortunately, the json_hint is for
deserializing the values of the tables to JSON.

Now since we're using an inflated column, we can't simply use
get_column() in TO_JSON anymore, because it will return the deflated
(raw) version of the database value.

So I've added another key for json_hint to designate inflated columns,
which then get properly serialized back into JSON.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>