Controller/API: Fix triggering jobsets
This was the only thing left which was still using jobset_input_alts.

We now use the ->> operator to do a lookup on a specific property. I've
left out the match for jobset_inputs.type = 'git' in the non-GitHub
version, because it was matching "value" from jobset_input_alts before
and thus it stays generic enough to trigger jobset for non-Git SCMs.

I've also added a status_ok() in the GitHub version, because otherwise
we get an error page as the reply, even though the jobsets were
triggered.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>