Jobset: Add a new property type "secret"
This property type is for values that only get exposed to the project
owner and are not shown via the "Configuration" tab of the jobset or the
API.

An example for this would be a token for private Git repositories.

The implementation could be more DRY in that we have a mapProperties
function that could be used for validateProperties as well, but we'd
have to restructure the whole validation functions again.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>