Jobset: Don't display properties without a spec
While in general this wouldn't be a problem, because the values would be
sent out as-is using the API it will become a problem as soon as type
"secret" is used in a property and that property is no longer existing
in the spec.

It would then spit out all the secrets where there is no specification
for the type, which is definitely not what we want.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>